refactor: move util fuctions to files, update readme
