+++ .gitignore
@@ -1,5 +1,4 @@
 node_modules
 npm-debug.log
 .DS_Store
-cjs
 .v8flags*.json
